Create frequencyQueries.js
You are given  queries. Each query is of the form two integers described below:
-  : Insert x in your data structure.
-  : Delete one occurence of y from your data structure, if present.
-  : Check if any integer is present whose frequency is exactly . If yes, print 1 else 0.

The queries are given in the form of a 2-D array  of size  where  contains the operation, and  contains the data element. For example, you are given array: 
[[1,1],[2,2],[3,2],[1,1],[1,1],[2,1],[3,2]]

The results of each operation are:
Operation   Array       Output
(1,1)               [1]
(2,2)              [1]
(3,2)                                 0
(1,1)              [1,1]
(1,1)             [1,1,1]
(2,1)             [1,1]
(3,2)                                 1


expected output [0,1]