Create iceCreamParlor.js
Each time Sunny and Johnny take a trip to the Ice Cream Parlor, they pool their money to buy ice cream. On any given day, the parlor offers a line of flavors. Each flavor has a cost associated with it.

Given the value of MONEY and the COST of each flavor for t trips to the Ice Cream Parlor, help Sunny and Johnny choose two distinct flavors such that they spend their entire pool of money during each visit. ID numbers are the 1- based index number associated with a COST. For each trip to the parlor, print the ID numbers for the two types of ice cream that Sunny and Johnny purchase as two space-separated integers on a new line. You must print the smaller ID first and the larger ID second.

For example, there are n = 5 flavors having COST=[2,1,3,5,6]. Together they have MONEY = 5 to spend. They would purchase flavor ID's 1 and 3 for a cost of 2 + 3 = 5. Use 1 based indexing for your response.
SAMPLE INPUT
2
4
5
1 4 5 3 2
4
4
2 2 4 3


SAMPLE OUTPUT
1 4
1 2